implemented required changes
